services/google_workspace_client/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
